chore: adjust ChatInput icons
