--- apps/mobile/src/features/chat/ChatInput/components/IconBtn/styles.ts
@@ -1,28 +1,0 @@
-import { ViewStyle } from 'react-native';
-
-import { createStyles } from '@/theme';
-
-const PADDING_SIZE = 16;
-
-export const useStyles = createStyles((token) => ({
-  buttonBase: {
-    alignItems: 'center',
-    backgroundColor: token.colorBgElevated,
-    borderColor: token.colorBorder,
-    borderRadius: token.borderRadiusLG + PADDING_SIZE,
-    borderWidth: 1,
-    justifyContent: 'center',
-    paddingHorizontal: token.paddingXS,
-  } as ViewStyle,
-
-  buttonPrimary: {
-    backgroundColor: token.colorPrimary,
-    borderColor: token.colorBgElevated,
-    color: token.colorTextLightSolid,
-  } as ViewStyle,
-}));
-
-export const getButtonSize = (size: number) => ({
-  height: size + PADDING_SIZE,
-  width: size + PADDING_SIZE,
-});
