+++ apps/mobile/src/features/chat/actions/NewChatBtn/index.tsx
@@ -1,12 +1,10 @@
 import { MessageSquarePlus } from 'lucide-react-native';
-import { useThemeToken } from '@/theme';
 import { useCallback } from 'react';
-import { ICON_SIZE } from '@/const/common';
-import IconBtn from '../IconBtn';
+import { ActionIcon } from '@/components';
 import { useSwitchTopic } from '@/hooks/useSwitchSession';
+import { ICON_SIZE_LARGE } from '@/const/common';
 
 const NewChatBtn = () => {
-  const token = useThemeToken();
   const switchTopic = useSwitchTopic();
 
   const handleNewChatTopic = useCallback(() => {
@@ -14,10 +12,7 @@ const NewChatBtn = () => {
   }, [switchTopic]);
 
   return (
-    <IconBtn
-      icon={<MessageSquarePlus color={token.colorText} size={ICON_SIZE} />}
-      onPress={handleNewChatTopic}
-    />
+    <ActionIcon icon={MessageSquarePlus} onPress={handleNewChatTopic} size={ICON_SIZE_LARGE} />
   );
 };
