+++ apps/mobile/src/features/chat/actions/SenderBtn/StopLoadingIcon/index.tsx
