+++ apps/mobile/src/features/chat/actions/SenderBtn/StopLoadingIcon/style.ts
